refactor[#470] : 404 페이지 파일 이름 수정
